+++ src/main/java/net/oneandone/pommes/mount/Action.java
@@ -21,14 +21,12 @@ import net.oneandone.pommes.scm.Scm;
 import net.oneandone.sushi.fs.file.FileNode;
 
 public abstract class Action implements Comparable<Action> {
-    public final Scm scm;
     public final FileNode directory;
-    public final Gav artifact;
+    public final String extra;
 
-    protected Action(Scm scm, FileNode directory, Gav artifact) {
-        this.scm = scm;
+    protected Action(FileNode directory, String extra) {
         this.directory = directory;
-        this.artifact = artifact;
+        this.extra = extra;
     }
 
     public abstract char status();
@@ -36,7 +34,7 @@ public abstract class Action implements Comparable<Action> {
     public abstract void run(Console console) throws Exception;
 
     public String toString() {
-        return status() + " " + directory + " (" + artifact + ")";
+        return status() + " " + directory.toString() + " " + extra;
     }
 
     @Override